to_do_django/requirements.txt
asgiref==3.7.2
attrs==23.1.0
Django==5.0
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.27.0
gunicorn==21.2.0
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
packaging==23.2
psycopg2==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
rpds-py==0.15.2
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
